Work around dropped mouse movement events on the world map.
I haven't traced the root cause of the problem, but not all mouse movement
events received by the engine core are sent to the world map renderer.  As
a result, the world map's scrolling tends to lag behind the mouse when it
is moved quickly across the screen.  This patch works around the problem
by adding an independent "last position" vector to WorldMapRender, and
using the difference between that and the current mouse position to
compute the mouse movement vector.