Adjust the zoom factor geometrically rather than arithmetically.
This patch changes the operation of the mouse scroll wheel to multiply or
divide the current zoom factor by a set amount, rather than adding or
subtracting; so, for example, instead of zooming in from 1.0 -> 1.2 -> 1.4
-> 1.6, the new code zooms from 1.0 -> 1.2 -> 1.44 -> 1.728, such that each
step is 20% of the current zoom value (and conversely for zooming out).
This gives a smoother feel to zooming throughout the entire scale range.