PSP port 2/40: Add basic PSP changes to the core code under BBGE_BUILD_PSP.
This is just the base set of changes needed to get the program to
interface with the PSP-specific code and build properly, including
things like the definition of main() and other system-specific routines.
(Note that it won't run on a real PSP yet, because it uses too much
memory during startup.)  All PSP-specific code in the Aquaria/ and BBGE/
directories is predicated on #define BBGE_BUILD_PSP.

This patch also includes a few fixes for SDL code that was not protected
by #ifdef BBGE_BUILD_SDL, and a tweak to the texture size calculation in
BBGE/AfterEffect.cpp to recognize screen sizes smaller than 512x512
(though aftereffects aren't currently implemented on the PSP due to a
lack of framebuffer object support).

There are a few FIXME notes scattered throughout this and subsequent
patches on things I could not confirm about the behavior of the current
code; please feel free to comment on or suggest changes to these.