PSP port 6/40: Use the directional pad for menu cursor movement.
This patch enables the use of the PSP's directional pad for moving the
menu cursor in all menus.  (It makes use of the X360_DPAD_* buttons
defined for the Xbox 360 controller, which are renamed to JOY1_DPAD_* by
this patch.)

For user interface consistency, menu control using the the analog pad is
_disabled_ by this patch, so that the analog pad is always used for
movement (swimming, etc.) and the directional pad is always used for
menu item selection.  This behavior is only applied when building for
the PSP; joystick control for other platforms is left unchanged.