PSP port 8/40: Avoid all filesystem writes on the PSP.
This patch disables all code which can write to the filesystem when
building for the PSP, in keeping with the PSP paradigm of a read-only
filesystem with writes handled exclusively through specialized OS
interfaces (the save data utility, for example).  Debug log messages
are directed through the same interface used for the DMSG() macro in
PSP code (which conveniently also prepends a timestamp to each line),
and all functions which write to temporary files are #ifdef'd out,
since none of them are called when running on the PSP anyway.

One minor additional change is the addition of a "const char *" version
of debugLog(), which is slightly more efficient than the std::string
version when logging literal text.