PSP port 11/40: Implement and begin using a lightweight string parser.
This patch adds the SimpleIStringStream class, which (as its name
implies) is a simpler version of std::istringstream from the STL.
Whether because of GCC's particular implementation or just due to the
inherent complexity of STL I'm not certain, but std::istringstream is
more than an order of magnitude slower than the equivalent code in
plain C for some operations.  This much simpler version is only 5-10%
slower than plain C code.

At the moment I'm only using it in the particle data loader, where
STL's inefficiency was most noticeable, but it could probably stand to
be used more widely in the code.