PSP port 9/40: Add a readFile() function to read entire files at once.
This patch adds the global function readFile(), which can be used to
read an entire file into a memory buffer.  On the PSP, this can be
significantly more efficient than using stdio or C++ streams to read
the file, since there is no disk cache or readahead functionality to
offset the overhead of repeated system calls.