PSP port 10/40: Add an icon to system save files, if available.
This patch adds code to load the file ICON0.PNG from the game's
installation directory (or the package file, if present there) and
use it as the icon for the user settings and achievements save files.
If the file cannot be loaded, no icon is stored with the save files.