PSP port 14/40: Add a closeLuaVM() method to parallel openLuaVM().
This patch adds the ScriptInterface::closeLuaVM() method as a parallel
to openLuaVM(), and replaces all calls to lua_close() with calls to this
new method.  At the moment, closeLuaVM() just calls lua_close() itself,
so there's no change in behavior, but this level of abstraction could be
useful in the future, e.g. if scripts are changed to run in Lua threads.