PSP port 17/40: double -> float conversion, part 2 of 4 (variables/fields).
This patch changes variables and structure/class fields of type double
into type float where doing so would not have any visible effect (for
example, when the variable is passed to another routine which only
accepts a float).  Nearly all of the "double" variables in the program
fall into this category; however, I wasn't certain whether double
precision was needed in the homingIncr and homingness fields of
ShotData, so I left those alone.