PSP port 20/40: Remove unnecessary fields from class Vector.
This patch removes the v[] and v4[] arrays from class Vector which were
used for returning values from the getv() and getv4() methods
respectively.  This reduces sizeof(Vector) by 70%, from 40 to 12 bytes,
which in turn brings the size of higher-level structures down
drastically: for example, sizeof(Quad) falls from 6080 to 3640 with
this path, a reduction of 40%.

With this patch, the game is able to proceed past the intro sequence to
Naija's cave, with a good 5.5MB free after the scene is loaded.