PSP port 22/40: Compress the EventPtr and InterpolatedVector structures.
This patch comments out the unused field InterpolatedVector.fakeTimePassed,
changes the type of the EventPTr.evm field to unsigned char (since it
never takes on larger values than 255), and rearranges fields in the
EventPtr and InterpolatedVector structures to pack small fields together
so as not to waste bytes on padding.  The end result is a reduction in
memory usage of about 1.5MB in Naija's cave.