PSP port 27/40: Compress the RenderObject and Quad structures.
This patch comments out several unused (or effectively unused, e.g.
because they are set to a default value and never changed) fields in the
RenderObject and Quad structures to reduce memory usage, and rearranges
other fields to minimize the amount of internal padding required.  This
saves about 250k of memory in the Secret Cave Passage map.