PSP port 33/40: Don't display "Game Saved" until after the game is saved.
This patch delays the display of the "Game Saved" message until after
the game has actually been saved, which can take 1-2 seconds on the PSP.
As a consequence, the screenshot capture is now performed before opening
the save slot menu, rather than after closing it.  (This change applies
to all systems, not just the PSP.)

The patch also adds new core methods, Core::grabScreenshot() and
Core::grabCenteredScreenshot(), to avoid duplication of the code for
capturing a screenshot.  It could potentially be used with screen
transitions as well, though this patch doesn't go that far.