PSP port 32/40: Fix rendering of large repeated textures.
This patch adds a PSP-specific routine for rendering quads with repeated
textures that subdivides the quads into smaller units if they are larger
than what the hardware can handle.

As noted in the comments, the PSP's graphics hardware will hard-clip any
polygon with a vertex that lies outside the 4096x4096 pixel region
surrounding the center of the screen (or any arbitrarily selected point,
but the PSP engine always uses the center of the screen).  Since some
repeating background textures cover areas larger than this, they have to
be split up into smaller subquads in order to be rendered properly.