PSP port 31/40: Fix rendering of full-screen quads.
This patch fixes the rendering of full-screen quads on the PSP so that
they don't miss the rightmost pixel column on the screen.  I suspect the
cause of this is similar to the screen transition texture offset, i.e.
rounding error in converting to and from the virtual screen size.