PSP port 34/40: Adjust save file screenshot zoom to match the PSP's display.
This patch adjusts the "zoom" transition applied to save screenshots
when starting the game to fill the entire PSP screen, not just the 4:3
aspect ratio portion.  The texture coordinates are also adjusted so the
screenshot does not appear "stretched" (note that PSP save file
screenshots are 144x80, rather than the 4:3 aspect ratio used on PCs).