Run all Lua scripts as independent threads in the same Lua state.
This patch changes the behavior of ScriptInterface::initLuaVM() and
closeLuaVM() to create and destroy Lua threads within a single Lua state
rather than create and destroy independent states, and establishes a
base state in which those threads are created.

Since the current scripts are not designed to be run as threads, the
code provides each thread with an independent environment for global
variables.  This thread environment includes a fallback lookup to the
base state's global table (by linking it to the new environment's
__index metamethod), allowing the thread to access all functions defined
by the base Lua libraries as well as the Aquaria-specific functions
registered in the base state.  (This is admittedly something of a hack,
and ideally the scripts should be rewritten to use local variables
rather than globals.)

This change reduces Lua's memory usage by up to 80% in some cases, for
example:

    Title screen:  3552008 ->  923406 bytes
   Training cave: 13128594 -> 2742677 bytes
     Home Waters: 21100917 -> 4715006 bytes

It also modestly reduces (by about 10%) the loading time for new areas.
While the nested thread environment does require an extra table lookup
for every function called from Lua, I have not observed any significant
increase in CPU usage while playing.