Use a bitmap instead of a list for tracking visited areas of the map.
This patch changes the WorldMapTile class to use a bitmap rather than a
list of IntPairs for tracking which areas of the map have been visited
(the MAPVIS feature).  This reduces worst-case memory usage (all areas
of all map tiles visited) by a factor of 128.

This change also resolves a bug in the current code which causes the
visited lists to grow in size every time the game is saved, since the
lists are not cleared at that time.  As a result, uncompressed save data
sizes could grow by an order of magnitude or more over the minimum
necessary size, depending on how frequently the player saved their game.
(This would in turn cause the game to freeze for a short time while
displaying a save slot page with one or more of these ballooned files.)
Any save files suffering from this bug can be fixed by loading and then
saving the game after applying this patch; a typical compressed save
file (*.aqs) from the endgame should be around 40-50k.

This patch does _not_ alter the save data format itself, only the method
used to store the data in memory.