Add an alternate format for visited area data in the save file.
This patch adds an alternate format for the visited-area (MAPVIS) data
stored in save files.  Rather than a textual list of coordinate pairs,
this alternate format writes the data as a base64-encoded string
containing the raw bitmap data stored for each world map tile.  This
has the advantage of only requiring a small, fixed amount of memory to
decode, and can be parsed much more quickly than reading and converting
a long stream of numbers from a string.

The alternate format is enabled by defining the preprocessor macro
AQUARIA_SAVE_MAPVIS_RAW, which is undefined by default.

IMPORTANT:  As noted in the DSQ.h comments, save files generated with
this feature enabled WILL BE INCOMPATIBLE with builds of Aquaria which
do not include this patch!  This feature is intended primarily for
portable systems and other platforms where storage space or memory are
at such a premium that the current text-based format (potentially up to
1.3MB, with three copies required in memory) is impractical.