Fix OpenGL implementations of Texture::{read,write}().
This patch adds the missing implementation of Texture::read() for OpenGL,
and corrects the Texture::write() implementation so it only writes the
pixels once.  (It looks like the original intent was to write to all
mipmap levels, but that would require manually resizing the input data
and other graphical contortions, so I've changed it to just write to the
base texture data.)

These methods are not used by the current Aquaria code, but I expect to
reference them in a future patch.