Implement alternative visited-area display logic for the PSP.
This patch removes the unused VIS_PARTICLES and VIS_COPY display methods
in WorldMapRender.cpp, and replaces them with the new VIS_WRITE method.
This method writes alpha values for visited and unvisited areas directly
to the pixel data for the current map tile's texture, restoring the
original texture data from a backup buffer when switching to a different
tile.  For simplicity, the code assumes that the world map tile textures
have been converted to indexed-color format using the "pngtotex" tool,
and will refuse to work on direct-color textures loaded from PNG files.

Generic code using Texture::read() and Texture::write() is also included
in this patch, so that VIS_WRITE should function on any system.
However, the hard boundary lines generated by this method may not look
good on high-resolution screens; in addition, the current implementation
of Texture::write() does not update mipmaps, so the visited/unvisited
distinction will disappear as the player zooms out on the map.

With this patch, the PSP is capable of rendering the visited-area effect
safely, so AQUARIA_BUILD_MAPVIS is once more enabled for PSP builds.