Fix regression causing crashes on PCs when changing resolution.
This patch fixes a regression introduced in PSP import changeset #12/40
(r36:3ad4a420d7be), which caused the program to often crash on PCs when
changing the screen resolution or switching between windowed and
fullscreen mode.  The regression was caused by an incorrect assumption
about calling patterns in the code, specifically that DSQ::destroyFonts()
would never be called twice in succession.