Invert up/down and left/right joystick scrolling in the world map.
This patch inverts the direction of scrolling in the world map when
using joystick control, so that (for example) pressing up on the stick
will cause the map to scroll down.  In other words, with this patch the
world map acts as though the joystick is moving the cursor -- which is
visible, though fixed in the center of the screen -- rather than moving
the map itself.  (I kept finding myself scrolling in the wrong direction
by accident, so I think this way may be more natural, but comments
either for or against are welcome.)