Remember the last save slot used for saving or loading a game.
This patch adds a new user preference field, "saveSlot", which records
the index of the most recent slot in which a game was saved or loaded
(starting from 0).  That slot is then used as the default when opening
the load or save menu, so that the cursor is automatically moved to it
when using keyboard or joystick control.  (The cursor is not moved when
using mouse control, since the actual mouse position is given priority,
but the most recently used save slot is still updated and recorded in
the preferences file.)

As a side effect, this patch also fixes a crash when reading the
preferences file if certain XML attributes were missing.