Add a "drop food" action for use with keyboard/joystick control.
This patch adds a new "drop selected food" action for use in the food
menu, allowing food to be dropped into the environment when using
keyboard or joystick control.  No keys or buttons are assigned to this
action by default, but they can be set through the Key Config menu in
the same manner as other actions.

Note that at present, there is no "confirm" step for dropping food; the
item simply disappears from the food list, and will appear in the game
world when the menu is closed.  I'm still thinking about how best to
implement confirmation while also behaving well if mouse control is used
at the same time.

As a side effect, this patch also avoids a (deliberate) crash when one
or more standard actions are missing from the user preferences file.
(This would be impossible with the commercial distribution unless the
user edited the preferences file in a text editor, but it arises with
the addition of this action, since the new action won't be present in
existing preferences files.)