Allow food to be dropped using "Cook" + "Food Left" on the PSP.
This patch allows food to be dropped on the PSP by holding the "Cook"
button (default Select) and pressing the "Food Left" button (default
Triangle) while all cooking slots are empty.

This is admittedly a rather ugly hack, but it's necessitated by the
limited number of buttons available on the PSP.  However, it might be
worked around in the future with a rethink of the overall control
interface, since currently the right-mouse-button equivalent, Circle,
serves the same purpose as Food Right (Square) in the food menu.