Render loop optimization 3/4: Devirtualize RenderObject::getCullRadius() and isOnScreen().
This patch changes the RenderObject::getCullRadius() and isOnScreen()
virtual methods into ordinary (non-virtual) class methods, allowing
them to be inlined and giving around a 5% decrease in time taken to
execute the render loop.

The inlined version of RenderObject::isOnScreen() added in this patch
omits the alpha.x==0, parent!=0, and cull==false tests from the original
code, since those are tested in Core::render() prior to calling
RenderObject::isOnScreen() (and there are no other callers of
isOnScreen()).

As a result of this change, the BoxElement class in
Aquaria/BoxElement.cpp can no longer override isOnScreen().  However,
BoxElements do not seem to be used in any of the actual data, so this is
presumed not to be a problem.