Minimap optimization 4/6: Refactor the warp/save/cooking point rendering loop.
This patch refactors the loop which renders warp, save, and cooking
points on the minimap, merging common code and skipping out of the loop
as soon as possible on paths for which nothing is rendered.  The patch
also moves the texRipple->apply() call out of the loop, on the
assumption that there will usually be either zero or one cooking points
and many more warp/save points.