Minimap optimization 5/6: Refactor the health bar rendering loops.
This patch slightly refactors the loops used to draw the health bar
(circle), pulling a few GL calls out of the loops where they don't need
to be called repeatedly.  The patch also changes the coordinate
calculations to follow the mathematical standard of x=cos(a), y=sin(a),
with the Y coordinate inverted because screen coordinates increase going
downward (instead of upward as in the Cartesian plane).