Minimap optimization 6/6: Use lookup tables to avoid sin/cos operations.
This patch adds lookup tables for use by the minimap and health bar
rendering loops, avoiding the need to repeatedly call the sinf() and
cosf() functions and cutting rendering time by more than half (as
measured on the PSP).

As a side effect of this patch, the position of the "maximum health"
marker will be rounded to one of the 64 lookup table coordinate pairs.
This was an intentional decision, as the potential error (<10%) is small
enough that the player won't be confused, and this method ensures that
there will always be a health "bit" directly underneath the marker when
Naija's health is full.