Increase minimap display size and decrease map resolution for the PSP.
This patch increases the displayed size of the minimap on the PSP so it
is large enough to be easily viewable by the player, and decreases the
resolution of the minimap itself (i.e., increases the distance between
obstruction grid taps) to further reduce rendering time.  This latter
change has the potential to alter the visibility of narrow passages or
obstructions, though I haven't yet seen any cases where it causes a
problem.

As a side effect of this patch, the minimap now moves closer to the top
of the screen when the in-game menu is opened.