Keep linked lists of map elements on each element layer.
This patch adds a bgLayerNext field to the Element class and uses it to
link elements on each element layer in a list specific to that layer.
Elements on a particular layer can then be iterated by getting the first
element from DSQ::getFirstElementOnLayer() and following the bgLayerNext
pointers until a NULL is reached.

This patch also changes the Entity::isInDarkness() method to use these
lists, reducing execution time by a factor of roughly 1,000 (1 msec ->
1 usec, as measured on the PSP) in the Home Waters area.

To ensure that the lists are properly generated, the "elements" field of
the DSQ class has been marked protected, and all code which formerly
accessed this field has been changed to use accessor functions instead.
New accessor functions getNumElements(), getElement(), and
getElementsCopy() have been added for this purpose.