Adjust joystick input handling to give finer-grained control over speed.
This patch tweaks the joystick input handling in Aquaria/Avatar.cpp to
provide a wider range of "slow swim" input, allowing finer-grained
control of Naija when using the PSP's analog pad (which is not well
suited to fine control).  The input threshold values are parameterized
so they can be readjusted later or for other systems as necessary; they
could potentially be turned into user settings as well.