Keep linked lists of paths of each path type.
This patch adds a nextOfType field to the Path class and uses it to link
paths of each distinct path type in a list specific to that type,
similar to the changes implemented for elements in r136:a97148c3c017.
This change reduces execution time for getNearestPath(Vector,pathType)
and other type-based path iteration loops by a factor of 10 in areas
with many paths (such as Home Waters).

As with the elements change, the "paths" field of the Game class has
been marked protected.