Piscine education 2/5: Avoid O(N^2) complexity in flock state calculation.
This patch avoids some O(N^2) loops in the calculation of flock state
(center and heading) and halves the number of checks needed to find the
nearest neighbor to each FlockEntity, further reducing the time required
to update FlockEntities.  As a side effect, this patch also ensures that
each FlockEntity in a flock sees a consistent flock state in its update
call, which may slightly improve the naturalness of entity movement.