Piscine education 3/5: Optimize obstruction check.
This patch optimizes the obstruction-checking loop in
SchoolFish::applyAvoidance() to avoid unnecessary vector calculations.
The patch also removes the lastVel field from the Entity class, since it
is only used by SchoolFish entities, and instead stores the magnitude of
the previous velocity in the new SchoolFish.lastSpeed field.

The TileVector.h hunk of the patch is only to add "const" to the
TileVector methods so they can be used with a const instance.