Piscine education 4/5: Reduce obstruction check granularity on the PSP.
This patch reduces the granularity of obstruction checking in
SchoolFish::applyAvoidance() by a factor of 9 when running on the PSP,
so that only every third tile location in each direction (thus 1 out of
every 9 tiles) is checked.  This provides a significant (>1ms) reduction
in execution time, at the potential cost of fish getting stuck on small
obstacles; as noted in the comments, an alternate method of managing
obstruction data may need to be devised if this proves to be a problem
in actual gameplay.