Fix missing-parentheses warnings.
This patch fixes several warnings about potential order-of-operation
problems due to missing parentheses, none of which was an actual bug but
one of which (in the Lua interface function l_entity_getNearestEntity())
only worked correctly by coincidence.  Two "if" blocks in Game.cpp were
revealed to be redundant by these fixes and were deleted.