Use HUGE_VALF instead of -1 when finding smallest values.
This patch changes a number of "find the smallest value" loops to use
HUGE_VALF instead of -1 as the initial "smallest" value, thus saving a
comparison against -1 inside the loop.  For loops which also had a test
against an upper limit, the "smallest" value is initialized to that
limit instead of HUGE_VALF, thus saving a further comparison against
the limit.

In addition, some loops which previously used integers to hold the
current and smallest distances have been changed to use floats, since
it is typically less expensive to directly compare two floating-point
values than to convert from floating-point to integer and then compare
two integers.

This patch gives an overall 2-3% boost to execution speed as measured
on the PSP.