Fix OT_BLACK obstructions not being drawn at the bottom of a map.
This patch fixes a regression introduced in r177 (e4b0246ccd9f) which
caused OT_BLACK (solid black) obstructions not to be drawn in the
bottommost obstruction tile row of a map in GridRender::onRender().
This appears to be due to carrying over the +1/-1 OT_BLACK logic from
the earlier code, so that logic has been removed; however, the purpose
of the original logic is unclear, so this change may result in
graphical regressions elsewhere.