Lua interface cleanup 3/11: Move helper functions to the top of the file.
This patch rearranges function definitions so that all helper functions
(such as Lua parameter extractors) are located at the top of the file,
above the Lua interface functions.

The object code is confirmed not to change with this patch, with the
exception of code movement and corresponding internal symbol changes
resulting from the rearrangement of the source code.  (The individual
functions are each identical at the machine code level.)