Lua interface cleanup 5/11: Use luaReturn*() macros to return values to Lua.
This patch changes all returns from Lua interface functions to use the
luaReturn*() shortcut macros rather than manually pushing values onto
the Lua stack and returning.  The patch also adds new luaReturn*()
macros for other types of return values not previously handled:
integers, (non-integer) numbers, strings, and 2- or 3-element vectors.

The object code is confirmed not to change with this patch, with the
exception that (in a PSP build) the order of initialization of two
registers in l_entity_getNearestBoneToPosition() is reversed; this
change has no visible effect.