Lua interface cleanup 11/11: Merge/rename duplicate Lua function names.
This patch renames several Lua interface functions to avoid having
multiple names for the same function (or in the case of "inp", to use a
clearer function name):

   avatar_setHeadTexture --> setNaijaHeadTexture
   bone_color -------------> bone_setColor
   bone_getidx ------------> bone_getIndex
   entity_flipTo ----------> entity_flipToEntity
   getEntityByName --------> getEntity
   getNodeByName ----------> getNode
   inp --------------------> toggleInput
   isPlayingVoice ---------> isStreamingVoice
   playVoice --------------> voice
   setOverrideCullRadius --> setCullRadius

This patch does not change the duplicate names "entity_color" (for
"entity_setColor"), "entity_fh" (for "entity_flipHorizontal"), or
"entity_isfh" (for "entity_isFlippedHorizontal") to avoid the risk of a
regression, since both names for each of these functions are used widely
throughout the game scripts.

IMPORTANT NOTE: This patch BREAKS COMPATIBILITY with existing scripts.
The game scripts included in the source code are updated by this patch
to function properly, but any mod scripts or other custom scripts which
make use of the changed names will not work until they are updated.
(Except for "toggleInput", the new names will also work with the
original Aquaria executable.  An "inp" alias can be added back to this
source code if that functionality turns out to be needed by any mods.)