Use the native strcasecmp() to implement nocasecmp() where possible.
This patch replaces the custom nocasecmp() string comparison function
with an inline function which calls strcasecmp() on platforms where the
latter is available, thus allowing the compiler to optimize the compare
when possible.  The patch currently assumes the function is available on
Unix platforms and on the PSP, but another possibility would be to add a
specific conditional (e.g. HAS_STRCASECMP) and define that for non-PSP
platforms based on a test in the CMake control file.

This patch regrettably does not address the POSIX misfeature of naming
the case-_insensitive_ compare function as though it is case-_sensitive_.