Rename NUM_PATH_TYPES to PATH_MAX.
This patch renames the NUM_PATH_TYPES constant added in r144
(bf7482016686) to PATH_MAX and makes it an enumeration constant instead
of a #define, to match the style of other enumerations in the code.

This use of PATH_MAX conflicts with the PATH_MAX constant defined in
POSIX system headers, but since there is no platform-specific code in
the Aquaria subdirectory that makes use of this constant, the system
definition (if any) is undefined in Path.h.