Fix a logic bug causing incorrect entity IDs to be generated.
This patch fixes a regression introduced in r201 (32dda929a9bd) which
caused incorrect entity IDs to be generated when an ID collision was
detected while loading a map.  Among other things, this caused two of
the Mithalas Cathedral mid-bosses to regenerate on loading a saved game
in which they had been defeated.

Note that any games saved with r201 applied but without this fix may
contain incorrect data, potentially preventing the game from being
completed.