New script execution model, version 1 (Lua script changes).
This patch modifies all game scripts so that they function properly
under the new instanced script execution model.  This includes the
following changes:

- All scripts now include a "v = getVars()" line at the top of the
  script to obtain a reference to the instance-scope variable table.

- Variables which were confirmed to be unused have been deleted.

- Variables which were confirmed to be used only within the block in
  which they are defined have been localized to that block.

- All other variables have been moved to the instance-scope variable
  table by prepending "v." to every use of the variable.  However,
  obvious constants such as state IDs have been left as global.

This patch is intended to not change any script behavior; however, it is
possible that regressions have slipped in, most likely as a result of
prepending "v." to function parameters or other variables which should
not be treated as instance-scope.  It is recommended to enable at least
the "complainOnUndefLocal" setting in ScriptInterface.cpp for the time
being, so that any accesses to undefined instance-scope variables are
reported as error messages.

Needless to say, these updated scripts WILL NOT FUNCTION with earlier
builds of the Aquaria executable.