Use C++ named constants in script constant table.
This patch modifies the script constant table added in the previous
changeset to (where possible) use named constants from C++ rather than
giving the constant's value literally.  This is confirmed not to change
any constant values other than a slight (<1e-7) loss of precision in the
value of FRAME_TIME.

As a side effect, this patch also adds a named CR_DEFAULT constant to
ScriptedEntity.h to replace the literal constant used in
ScriptedEntity::setupConversationEntity().