Add a GUI-based PSP build program.
This patch adds a GUI-based program allowing the PSP port to be built
from the data set of the commercial release or Aquaria.  The compiled
game program and script files are linked directly into the executable
and output as part of the build process.

The GUI is implemented using GTK+, and should run on any platform
supported by GTK+ with version 2.16 or later of the libraries.
(However, I haven't yet succeeded in building a functioning Windows
executable.)