Allow songs to be checked on the song menu using joystick/keyboard controls.
This patch enables the use of the menu directional controls (e.g. arrow
keys or joystick directional pad) to highlight song icons in the song
menu, so the user is not required to use the mouse to do so.

(Judging from commented-out code, this seems to have been possible at
one point during development but was subsequently disabled for unknown
reasons.  However, since the song menu is currently unusable without a
mouse, I've taken the liberty of re-enabling it.  The code looks correct
on initial review, but please report any regressions.)