Import save menu fixes from PSP branch.
This patch imports changes to save menu behavior introduced on the PSP
branch in r57:6f8d97f62d98 (moving the screen capture to before opening
the save menu, and delaying display of the "Game Saved" message until
the game has actually been saved), as well as the fix in r88:8833dd3698f4
for a regression affecting save file screenshots.  As a result of this,
there should be slightly less of an overall delay when saving.