Implement on-the-fly decoding for Ogg Vorbis audio streams.
This patch implements on-the-fly decoding for Ogg Vorbis audio data
streamed from disk (music and voice files), notably eliminating the
short freezes in gameplay when voice playback starts at the beginning
of a cutscene, and also reducing memory usage by around 50-80MB
depending on the music being played.

On-the-fly decode of memory-buffered audio is not yet implemented in
this patch, but as a temporary workaround, the constant
STREAM_EVERYTHING can be set to true to force all audio to be streamed
from disk.  This saves around 100MB of memory, but increases the chance
of "pops" in the audio due to disk access delays.