Fix a few potential multithreading bugs and other minor code/documentation issues.
This patch corrects two potential race conditions in the file reading
manager, one of which is presumed to be the cause of rare crashes and
audio corruption, as well as a potential bug in the MP3 decoder which
could manifest depending on compiler optimizations (though current
versions of GCC seem to generate safe code).

The patch also includes a few miscellaneous tweaks to code style and the
README documentation.