Avoid saving the user settings file if it hasn't changed.
This patch adds a PSP-specific check to UserSettings::save() which
checks the generated XML data against the last data loaded or saved from
the settings file and skips the save operation if the data is unchanged.
Notably, this eliminates a delay of about 1 second when transitioning
from the Bit Blot logo animation to the title screen.