Return the cursor to the current title menu node when a submenu is closed.
This patch updates the title menu processing so that when a submenu
(such as the options menu or quit confirmation dialog) is closed under
keyboard or joystick control, the mouse cursor returns to the location
of the associated title menu node.  The cursor is _not_ moved when under
real mouse control, since the sudden jump in position could confuse the
user.

This is functionally the same change made in PSP import changeset #40/40
(r64:6782f2dfb696), except that the logic is implemented in the Lua
scripts for the title menu nodes.  Accordingly, this patch depends on
the getInputMode() function introduced in r291 (d6bfaea5626e).