Use HUGE_VALF insted of -1 as the default value for "smallest" searches.
This patch is a backport of the following three changesets on the PSP
branch:
   - r137:5bfde80b64ce (reduce operation count in getNearestPath())
   - r142:f73da62f01b1 (use HUGE_VALF instead of arbitrary large value)
   - r175:b11e1738ab68 (use HUGE_VALF for other loops as well)

The net effect of these patches is to rewrite "find the smallest value"
or "find the closest object" loops, changing the following construct:

    float closest = -1;
    Object *closestObject = 0;
    for (Object *object = firstObject(); object; object = nextObject()) {
        float dist = distanceToObject(object);
        if (closest == -1 || dist < closest)
            closest = dist, closestObject = object;
    }

by initializing the distance to the currently closest object to infinity
instead of a negative value, which allows the "closest == -1" test to be
eliminated.  (The constant HUGE_VALF is normally defined in <math.h> as
a single-precision infinity, but a fallback is supplied in BBGE/Base.h
for environments which are missing the constant.)  Depending on the loop
structure and execution architecture, removing that single comparison
can reduce execution time by several percent.

As a further optimization, loops which include a test for a maximum
value (such as "find the closest object within 100 units") have that
test merged into the distance test by initializing the closest-distance
variable to that limit instead of HUGE_VALF.

In addition, some loops which previously used integers to hold the
current and smallest distances have been changed to use floats, since
it is typically less expensive to directly compare two floating-point
values than to convert from floating-point to integer and then compare
two integers.

This branch is based on the floating-point-fixes branch as merged into
the "patched" branch at r257 (88eca1bdc031).